installer: fix network interface detection
There are a few issues with the NIC detection code in
network_up_for_answerfile() that will surface if:
a) you are doing a network install
b) your nics are not detected before it sets
   NIC_LIST="$(find_physical_nics)"

This problem typically occurs on multi-homed systems.

This fix cleans up the logic such that we iterate
over currently known NICs, brings the interface up,
then checks for carrier signal.  If no carrier, move on
to the next known interface.  Once all interfaces have
been checked and fail carrier check, sleep for 5 seconds
and start over (redetecting for newly available interfaces).

If carrier signal is detected, use that NIC and configure it
accordingly.

OXT-287

Signed-off-by: Chris Patterson <pattersonc@ainfosec.com>