installer: Cleanup existing LVM layout.
Cleaning the MBR does not trigger the device-mapper to wipe the PV/VG/LV
layout and can cause loss of data if a VG has PE on different disks, one
of which being the target-disk of the OpenXT installation.
Instead try hard to relocate and/or wipe PV/VG/LV on the current disk.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1107