Set-password: Collect all passwords at once
Query for the Recovery and TPM passwords at the same time as the system
administrator password in a single "mixed form".  Empty passwords are
prohibited.  Max password length is 64 - this needs to be hardcoded for
dialog.

For non-interactive installations, files stored via store_own_key() and
store_recovery_key() are automagically used if present.  Plaintext
passwords can also be included in the answerfile inside
<recovery-password> and <tpm-owner-password> fields.  store_own_key()
and store_recovery_key() shell functions are available by sourcing
/usr/lib/openxt/key-functions.

OXT-1130

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>