Unlock-config: Verify recovery password before starting
Ensure we have the proper recovery password before we start modifying
the disk during a refresh.  This avoids clobbering the disk only to find
we cannot finish the installation during "Refresh".

If /config is already mounted, we can just continue.

If the recovery key is present, check and continue if good.

If the above fail and we are automated, we fail.

OXT-1130

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>