part1: Stash passwords and sanitize answerfile
We don't want plaintext passwords hanging around and ending up in
debugging info.  So stash the passwords into the keystore area as early
as possible and sanitize them out of the answerfile.

Instead of deleting the line, replace the tag value with the text
"SANITIZED".  This allows a post-mortem to differentiate whether or not
a tag was set.

We need to provide a default recovery password to make non-interactive
installations work by default.  Otherwise the installer bails out since
the password is required.

Enforce passwords having a length.  An empty TPM password, at least,
fails the reseal after installation.  An empty recovery password seems
to work, but we still disallow it.

OXT-1130

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>