TPM-Timeout: Clear attempted owner password
When we hit a timeout, we probably had the wrong owner password entered.
If we don't remove it, we return to Set-password, but it will
immediately exit.  TPM-Check will then immediately re-try the stored
password, which may just further increase the timeout.

Remove the file, so that Set-password will requery.

OXT-1033

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>