TPM-Check: Remove well known secret owner check
tpm_is_owner_wks checks if the TPM owner is the well-known secret.  It
does this by authenticating to the TPM.  If the TPM is already owned,
this adds another failure to the TPM.  Too many failures and you'll hit
the TPM's timeout.

Since we don't actually do anything with the result of tpm_is_owner_wks,
just remove the test.

OXT-1308

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>