install-main: Return failure from create_lv
When lvcreate fails we should bubble the error up to the caller.

During development, I had a huge debugging dom0 which left only 24MB
free in the volume group after being written to /dev/xenclient/root.new.
This wasn't enough for the 143MB upgrade-compat image, so its LV could
not be created.  The error was not caught, so execution continued until
install-bootloader eventually failed and failed the OTA.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>