[stable-9] Add boot VGA device warning
Add a check to the host capabilities section that tests whether the
onboard VGA device is the boot VGA device. If not the user is served with
a warning. User can continue installation if desired, they are instructed
to change primary display after the install succeeds in such an event.

Related to OXT-1600

Signed-off-by: Tyler McGavran <mcgavrant@ainfosec.com>