part2: Delay forward seal until VMs are not running
Don't run seal-system directly. The TPM interactions could have their MMIO data
snooped. Instead we touch a marker file and let the delayed-seal initscript
detect and run the forward seal at shutdown time. This is done when the VMs are
not running, so we shouldn't have anyone around to snoop on the transaction.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>