part2: Allow installing into /config
Currently, /config is not mounted under mount_dom0.  This means files
cannot be installed there while the installer runs.  However, /config is
an encrypted partition, so it is the convenient place to store files
needing offline protection.

The installer already has permission to mount /config.  mount_config
already exists to mount /config, but it's current use is limited to
specific locations where access is needed.  It's not clear if not
mounting /config was a design decision, or if it was simply never needed
before.  Maybe it was done done to protect against stray writes.
However there is no actual restriction on accessing /config.

While mount_dom0 could be expanded to always mount /config, take a more
conservative approach and just extended install_file() and commit_file()
to only mount /config if and when needed by checking for a destination
of "/config/".

/config will always be available for mounting - however, there is and
always has been, and implicit need for install_dom0()/upgrade_dom() to
run before other steps.
For install, install_dom0 created and luksOpen-ed /dev/mapper/config.
For OTA, /config is already mounted in the host.
For offline upgrade, the config recovery password is verified before
proceeding.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>